revert datajob stepfunctions to the old way of configuration
