+++ datajob/datajob.py
@@ -7,6 +7,7 @@ from pathlib import Path
 import typer
 from stepfunctions.workflow.widgets.utils import create_sfn_execution_url
 
+import datajob.stepfunctions.stepfunctions_execute
 from datajob import console
 from datajob import stepfunctions
 from datajob.package import wheel
@@ -108,10 +109,14 @@ def execute(
         ..., help="the full name of the state machine you want to execute."
     )
 ):
-    state_machine_arn = stepfunctions._find_state_machine_arn(state_machine)
+    state_machine_arn = (
+        datajob.stepfunctions.stepfunctions_execute._find_state_machine_arn(
+            state_machine
+        )
+    )
     console.log(f"executing: {state_machine}")
-    execution = stepfunctions._execute(state_machine_arn)
-    status = stepfunctions._get_status(execution)
+    execution = datajob.stepfunctions.stepfunctions_execute._execute(state_machine_arn)
+    status = datajob.stepfunctions.stepfunctions_execute._get_status(execution)
     console.log(f"status: {status}")
     url = create_sfn_execution_url(execution.execution_arn)
     console.log(f"view the execution on the AWS console:")
