+++ datajob_tests/datajob_cli_tests/test_datajob_execute.py
@@ -29,7 +29,7 @@ class TestDatajobExecute(unittest.TestCase):
         execution.configure_mock(**attrs)
         return execution
 
-    @patch("datajob.stepfunctions._execute")
+    @patch("datajob.stepfunctions.stepfunctions_execute._execute")
     @patch.object(Workflow, "list_workflows")
     def test_datajob_cli_run_with_no_errors(self, m_list_workflow, m_execute):
         some_state_machine = "some-statemachine-1"
