+++ examples/data_pipeline_pyspark/datajob_stack.py
@@ -4,7 +4,7 @@ from aws_cdk import core
 
 from datajob.datajob_stack import DataJobStack
 from datajob.glue.glue_job import GlueJob
-from datajob.stepfunctions import StepfunctionsWorkflow
+from datajob.stepfunctions.stepfunctions_workflow import StepfunctionsWorkflow
 
 current_dir = str(pathlib.Path(__file__).parent.absolute())
